fix voucher payment

+++ MeApp/PopUps/ShareVoucher/MAShareVaucherViewController.swift
@@ -30,14 +30,10 @@ class MAShareVaucherViewController: MABasePopUpViewController {
             }else{
                 categoryNames.append(category.name+", ")
             }
-            
         }
         categoriesLabel.text = "\(categoryNames!)"
         categoryNameLabel.text = voucher.found.organization.name
         productNameLabel.text = "€\(voucher.amount!)"
-        
-        
-        
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -51,7 +47,7 @@ class MAShareVaucherViewController: MABasePopUpViewController {
     
     @IBAction func checkAmount(_ sender: Any) {
         if amount.text != "" {
-            if let limit = self.voucher.amount, Int(amount.text!)! > limit{
+            if let limit = Double(self.voucher.amount), Double(amount.text!)! > limit{
                 amount.errorMessage = "You enter more biger price then amount!"
             }else{
                 amount.errorMessage = nil
@@ -64,13 +60,13 @@ class MAShareVaucherViewController: MABasePopUpViewController {
     
     @IBAction func send(_ sender: Any) {
         if amount.text != "" {
-            if let limit = self.voucher.amount, Int(amount.text!)! > limit{
+            if let limit = self.voucher.amount, Double(amount.text!)! > limit{
                 amount.errorMessage = "You enter more biger price then amount!"
             }else{
                 amount.errorMessage = nil
                 let parameters: Parameters = [
                                   "organization_id" : voucher.allowedOrganizations.first?.id!,
-                                  "amount" : Int(self.amount.text!)]
+                                  "amount" : Double(self.amount.text!)]
                 TransactionVoucherRequest.makeTransaction(parameters: parameters, identityAdress: voucher.address, completion: { (transaction, statusCode) in
                     if statusCode == 201{
                             self.dismiss(animated: true, completion: nil)
